--- MyMovies/Domain/Interfaces/Repositories/TVShowDetailsRepository.swift
@@ -1,15 +1,0 @@
-//
-//  TVShowDetailsRepository.swift
-//  TVToday
-//
-//  Created by Jeans Ruiz on 1/16/20.
-//  Copyright © 2020 Jeans. All rights reserved.
-//
-
-import Foundation
-import RxSwift
-
-protocol TVShowDetailsRepository {
-  
-  func fetchTVShowDetails(with showId: Int) -> Observable<TVShowDetailResult>
-}
